move predicates utilities to utils package and change name of scale test manager (#4160)
