+++ systemtests/src/main/java/io/enmasse/systemtest/utils/Predicates.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.systemtest.apiclients;
+package io.enmasse.systemtest.utils;
 
 import java.util.Arrays;
 import java.util.Collection;
