+++ systemtests/src/test/java/io/enmasse/systemtest/iot/shared/control/CommandAndControlTest.java
@@ -5,9 +5,9 @@
 package io.enmasse.systemtest.iot.shared.control;
 
 import static io.enmasse.systemtest.TestTag.ACCEPTANCE;
-import static io.enmasse.systemtest.apiclients.Predicates.is;
 import static io.enmasse.systemtest.iot.MessageType.COMMAND_RESPONSE;
 import static io.enmasse.systemtest.iot.MessageType.TELEMETRY;
+import static io.enmasse.systemtest.utils.Predicates.is;
 import static java.net.HttpURLConnection.HTTP_ACCEPTED;
 import static java.net.HttpURLConnection.HTTP_OK;
 import static org.hamcrest.CoreMatchers.anyOf;
